Fix import errors after refactoring
#time 3m
